nile_django/requirements.txt
amqp==5.3.1
asgiref==3.11.1
async-timeout==5.0.1
billiard==4.2.4
celery==5.6.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron-descriptor==1.4.5
cryptography==46.0.7
Django==4.2.30
django-allauth==65.14.3
django-celery-beat==2.9.0
django-cors-headers==4.9.0
django-htmx==1.26.0
django-tailwind-cli==4.2.4
django-timezone-field==7.2.1
django-typer==3.4.0
django_celery_results==2.6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
exceptiongroup==1.3.1
idna==3.11
kombu==5.6.2
narwhals==2.19.0
numpy==2.0.2
openpyxl==3.1.5
packaging==26.0
pandas==2.3.3
plotly==6.7.0
polars==1.36.1
polars-runtime-32==1.36.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.12.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
pytz==2026.1.post1
redis==7.0.1
requests==2.32.5
semver==3.0.4
shellingham==1.5.4
six==1.17.0
sqlparse==0.5.5
typer-slim==0.20.1
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
whitenoise==6.11.0
